requirements.txt
praw
tweepy
google-api-python-client
beautifulsoup4
requests
pandas
nltk
selenium
webdriver-manager
python-dotenv
typing_extensions==4.5.0
pandas
pymongo
langdetect
flair
bson
textblob